[COZY-46] refactor: 코딩 컨벤션에 따른 리팩토링
