+++ src/main/java/com/cozymate/cozymate_server/domain/chat/repository/ChatRepository.java
@@ -1,7 +1,7 @@
-package com.cozymate.cozymate_server.domain.chat;
+package com.cozymate.cozymate_server.domain.chat.repository;
 
+import com.cozymate.cozymate_server.domain.chat.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRepository extends JpaRepository<Chat, Long> {
-
 }
