+++ src/main/java/com/cozymate/cozymate_server/domain/chat/service/ChatQueryService.java
@@ -1,0 +1,11 @@
+package com.cozymate.cozymate_server.domain.chat.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ChatQueryService {
+}
